--- Assets/BoxColliderScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bb93759f3e34b40adba04af8ed471128
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
