--- Assets/MovementController.cs
@@ -1,16 +1,0 @@
-using UnityEngine;
-public class Movementcontroller : MonoBehaviour
-{
-    public bool direction_right = true;
-    void Update()
-    {
-        if(direction_right){
-            //go right
-            transform.Translate(6f * Time.deltaTime, 0f, 0f);
-        }
-        else{
-            //go left
-            transform.Translate(-6f * Time.deltaTime, 0f, 0f);
-        }
-    }
-}
