--- Assets/MovementController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6498a72d975c54bf6b25c6bc1f806b34
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
