+++ Assets/PlayerController.cs
@@ -6,6 +6,7 @@ using UnityEngine.InputSystem;
 public class PlayerController : MonoBehaviour
 {
     public float jumpForce = 400f; 
+    public bool direction_right = true;
     Rigidbody2D rb;
 
     // Start called on start
@@ -26,6 +27,34 @@ public class PlayerController : MonoBehaviour
     // Update is called once per frame
     void Update()
     {
-        
+
+    }
+}
+
+public class Movementcontroller : MonoBehaviour
+{
+    void Update()
+    {
+        if(direction_right){
+            //go right
+            transform.Translate(6f * Time.deltaTime, 0f, 0f);
+        }
+        else{
+            //go left
+            transform.Translate(-6f * Time.deltaTime, 0f, 0f);
+        }
+    }
+}
+
+public class BoxColliderScript : MonoBehaviour
+{
+    public Movementcontroller movementController;
+    Rigidbody2D wall = this.GetComponent<Rigidbody2D>();
+    private void OnTriggerEnter(Collider other)
+    {
+        if (other.CompareTag("Wall"))
+        {
+            movementController.direction_right = !direction_right; //changes player direction
+        }
     }
 }
